新ScoreSaber対応、及び直接プレイリスト化機能追加

+++ playlist_tool.rb
@@ -27,8 +27,6 @@ require 'csv'
 require 'json'
 require 'time'
 require 'base64'
-require 'net/http'
-require 'nokogiri'
 require 'vr/vruby'
 require 'vr/vrcontrol'
 require 'vr/vrcomctl'
@@ -36,112 +34,129 @@ require 'vr/vrddrop.rb'
 require 'vr/clipboard'
 require '_frm_playlist_tool.rb'
 
-SETTING_FILE = EXE_DIR + 'setting.json'
-SS_SERVER = 'scoresaber.com'
-SS_PATH   = '/u/'
-SOFT_VER  = '20191118'
-
+SETTING_FILE  = EXE_DIR + 'setting.json'
+SS_API        = 'https://scoresaber.com/api/player/'
+CURL_TIMEOUT  = 20
+SOFT_VER      = '20220116'
 
 ####文字コード変換処理####
-#SJIS → UTF-8変換
+#SJIS → UTF-8変換#
 def utf8cv(str)
-  if str.kind_of?(String)
-    return NKF.nkf('-w --ic=CP932 -m0 -x',str)
+  if str.kind_of?(String)                       #引数に渡された内容が文字列の場合のみ変換処理をする
+    return NKF.nkf('-w --ic=CP932 -m0 -x',str)  #NKFを使ってSJISをUTF-8に変換して返す
   else
-    return str
+    return str                                  #文字列以外の場合はそのまま返す
+  end
+end
+
+#UTF-8 → SJIS変換#
+def sjiscv(str)
+  if str.kind_of?(String)                       #引数に渡された内容が文字列の場合のみ変換処理をする
+    return NKF.nkf('-W --oc=CP932 -m0 -x',str)  #NKFを使ってUTF-8をSJISに変換して返す
+  else
+    return str                                  #文字列以外の場合はそのまま返す
   end
 end
 
 ####メイン変換処理####
 #ScoreSaber 取得処理
-def user_score_get(userid,limit_page = false,prgrssbar = false,read_static = false)
-  proxy_class = Net::HTTP
-  #proxy_class = Net::HTTP::Proxy('proxy_server.com', 8080)  #Proxyサーバを使う時
+def user_score_get(userid,api_limit,limit = false,recent = false,prgrssbar = false,read_static = false)
 
+  player_data = {}
   ss_data = []
   page = 1
-  maxpage = 1
+  total = 0
   player = ''
-  rankchart = ''
+  profilePicture = ''
+  if recent
+    sort = "recent"
+  else
+    sort = "top"
+  end
+  url = "#{SS_API}#{userid}/full"
+  puts "#{userid}:UserDataGet"
+  begin
+    player_data = JSON.parse(`curl.exe --connect-timeout #{CURL_TIMEOUT} #{url}`)
+    player = player_data['name']
+    profile_picture = player_data['profilePicture']
+    total = player_data['scoreStats']['totalPlayCount']
+  rescue
+    puts player_data
+    return false
+  end
+  if limit
+    limit = total if limit > total
+    if api_limit >= limit
+      maxpage = 1
+      api_limit = limit
+    else
+      maxpage = limit / api_limit
+      maxpage += 1 if (limit % api_limit) > 0
+    end
+  else
+    maxpage = total / api_limit
+    maxpage += 1 if (total % api_limit) > 0
+  end
   while page <= maxpage do
-    #ScoreSaberからユーザのスコアリストのHTMLソースを取得
-    res = proxy_class.start(SS_SERVER) {|h|
-      h.get("#{SS_PATH}#{userid}&page=#{page}&sort=2")
-    }
-    #NokogiriでHTMLをパースし情報取得
-    doc = Nokogiri::HTML(res.body,nil,'utf-8')
-    #プレイヤー名
-    player = doc.search('h5[@class="title is-5"]').text.strip
-    #ランクのチャートデータ
-    rankchart = '0'
-    doc.search('script').text.each_line do |line|
-      if line.strip =~ /^data: \[[\d,]+\],$/
-        rankchart = line.strip.sub(/data: \[/,'').sub(/\],/,'')
-      end
+    api_limit = limit % api_limit if limit && page == maxpage && (limit % api_limit) != 0
+    puts "User:#{player} ScoreDataGet (#{page}/#{maxpage})"
+    #ScoreSaberからユーザのスコアデータを取得
+    url = "#{SS_API}#{userid}/scores?limit=#{api_limit}&sort=#{sort}&page=#{page}"
+    begin
+      scores_data = JSON.parse(`curl.exe --connect-timeout #{CURL_TIMEOUT} #{url}`)
+      playerScores = scores_data['playerScores']
+    rescue
+      puts scores_data
+      return false
+    end
+    unless playerScores
+      puts scores_data
+      return
     end
     #スコアテーブル
-    doc.search('table[@class="ranking songs"]').each do |table|
-      tr = table.search("tr")
-      tr.each do |a|
-        #ランク
-        rank = a.search('th[@class="rank"]').text.strip.sub(/#/,'').gsub(/,/,'')
-        next if rank == '' || rank == 'Rank'
-        #曲名
-        song_nodeset = a.search('th[@class="song"]')
-        song = song_nodeset.search('span[@class="songTop pp"]').text.strip
-        #マッパー
-        mapper = song_nodeset.search('span[@class="songTop mapper"]').text.strip
-        #曲ID
-        id   = song_nodeset.at("img")["src"].sub(/^.+\/(\w+)\.png$/,'\1').strip
-        #譜面ID
-        uid  = song_nodeset.at("a")["href"].sub(/\/leaderboard\//,'').strip
-        #記録日時
-        time = song_nodeset.at('span[@class="songBottom time"]')["title"].strip
-        time = Time.parse(time).localtime.strftime("%Y/%m/%d %H:%M:%S")
-        #何日前
-        songBottom_time = song_nodeset.search('span[@class="songBottom time"]').text.strip
-        #曲pp
-        score_nodeset = a.search('th[@class="score"]')
-        pp   = score_nodeset.search('span[@class="scoreTop ppValue"]').text.strip
-        #取得pp
-        ppw  = score_nodeset.search('span[@class="scoreTop ppWeightedValue"]').text.strip.gsub(/\(/,'').gsub(/pp\)/,'')
-        #スコア
-        score= score_nodeset.search('span[@class="scoreBottom"]').text.strip.gsub(/,/,'')
-        ss_data.push [uid,id,rank,time,songBottom_time,pp,ppw,score,mapper,song]
+    playerScores.each do |score|
+      if score['score']
+        rank   = score['score']['rank']
+        time   = score['score']['timeSet']
+        time   = Time.parse(time).localtime.strftime("%Y/%m/%d %H:%M:%S")
+        pp     = score['score']['pp']
+        ppw    = pp * score['score']['weight']
+        base_score    = score['score']['baseScore']
+        modified_score = score['score']['modifiedScore']
+      else
+        next
       end
-    end
-    #最大ページ情報の取得
-    doc.search('ul[@class="pagination-list"]').each do |ul|
-      li = ul.search("li")
-      li.each do |a|
-        maxpage = a.text.strip.gsub(/,/,'').to_i
+      if score['leaderboard']
+        song   = score['leaderboard']['songName']
+        song_sub = score['leaderboard']['songSubName']
+        mapper = score['leaderboard']['levelAuthorName']
+        id     = score['leaderboard']['songHash']
+        uid    = score['leaderboard']['id']
+        max_score     = score['leaderboard']['maxScore']
+        if score['leaderboard']['difficulty']
+          difficulty = score['leaderboard']['difficulty']['difficultyRaw']
+        else
+          difficulty = ""
+        end
+      else
+        next
       end
+      ss_data.push [uid,id,rank,time,pp,ppw,base_score,modified_score,max_score,mapper,song,song_sub,difficulty]
     end
     #画面更新
     if read_static
       SWin::Application.doevents
-      if limit_page
-        endpage = limit_page
-      else
-        endpage = maxpage
-      end
-      read_static.caption = "#{page}/#{endpage} page read"
+      read_static.caption = "#{page}/#{maxpage} page read"
     end
+    prgrssbar.position = page * 100 / maxpage if prgrssbar
     page += 1
     return false if ss_data == []
-    #取得ページ制限
-    if limit_page
-      prgrssbar.position = page * 100 / limit_page if prgrssbar
-      break if page > limit_page
-    else
-      prgrssbar.position = page * 100 / maxpage if prgrssbar
-    end
   end
-  return [player,rankchart,ss_data]
+  return [player,profile_picture,ss_data]
 end
 
 #プレイリスト作成
-def playlist_convert(cb_text,name_output,image_file,description,title,author,namecol,hashcol)
+def playlist_convert(cb_text,name_output,image_file,description,title,author,namecol,hashcol,sjis = true)
   cb_array = []
   cb_text.each_line do |line|
     cb_array.push line.split("\t")
@@ -165,10 +180,49 @@ def playlist_convert(cb_text,name_output,image_file,description,title,author,nam
     end
   end
   bplist_data['songs'] = songs_array
-  bplist_data['playlistTitle'] = utf8cv(title)
-  bplist_data['playlistDescription'] = utf8cv(description)
-  bplist_data['playlistAuthor'] = utf8cv(author)
+  bplist_data['playlistTitle'] = sjis ? utf8cv(title) : title
+  bplist_data['playlistDescription'] = sjis ? utf8cv(description) : description
+  bplist_data['playlistAuthor'] = sjis ? utf8cv(author) : author
   if File.exist?(image_file)
+    image_data = File.open(image_file, "rb") {|f| f.read }
+    image_base64 = Base64.encode64(image_data).split().join()
+    if File.extname(image_file) =~ /\.png$/i
+      type = 'png'
+    else
+      type = 'undefined'
+    end
+    bplist_data['image'] = "data:image/#{type};base64,#{image_base64}"
+  else
+    bplist_data['image'] = '1'
+  end
+  return bplist_data
+end
+
+def playlist_convert2(title,profile_picture,ss_data)
+  bplist_data = {}
+  songs_array = []
+  ss_data.each do |songs|
+    song_data = {}
+    song_data['songName'] = songs[10]
+    song_data['levelAuthorName'] = songs[9]
+    song_data['hash'] = songs[1]
+    characteristic = songs[12].split('_')[2].sub(/(Solo|Party|Multiplayer)/,'')
+    diff_name      = songs[12].split('_')[1]
+    song_data['difficulties'] = [{'characteristic' => characteristic,'name' => diff_name}]
+    songs_array.push song_data
+  end
+  bplist_data['songs'] = songs_array
+  bplist_data['playlistTitle'] = title
+  ext = "jpg"
+  ext = $1 if profile_picture =~ /\.(\w+)$/
+  image_file = "#{EXE_DIR}temp.#{ext}"
+  puts "user:#{title} Profile Picture Get"
+  begin
+    `curl.exe --connect-timeout #{CURL_TIMEOUT} --output "#{image_file}" #{profile_picture}`
+  rescue
+    image_file = false
+  end
+  if image_file && File.exist?(image_file)
     image_data = File.open(image_file, "rb") {|f| f.read }
     image_base64 = Base64.encode64(image_data).split().join()
     if File.extname(image_file) =~ /\.png$/i
@@ -189,42 +243,113 @@ module Frm_main_form
   
   #フォーム起動時処理
   def self_created
+    if $Exerb
+      #アイコンの設定
+      extractIconA = Win32API.new('shell32','ExtractIconA','LPI','L')
+      myIconData = extractIconA.Call(0, "#{EXE_DIR}#{File.basename(MAIN_RB, '.*')}.exe", 0)
+      sendMessage(128, 0, myIconData)
+    end
+    self.x = 330
+    self.y = 70
+    @limit_checkBox.check true
+    @tab_checkBox.check true
+    @radio_top.check true
+    @nameoutput_checkBox.check true
     if File.exist?(SETTING_FILE)
       setting = JSON.parse(File.read(SETTING_FILE))
       @userid_edit.caption = setting['UserID']          if setting['UserID']
-      @limit_checkBox.check true                        if setting['LimitPage']
-      @limit_edit.caption = setting['LimitPage']        if setting['LimitPage']
-      @tab_checkBox.check true                          if setting['TSV']
+      @limit_checkBox.check false                       if setting['Limit'] == false
+      @limit_edit.caption = setting['Limit']            if setting['Limit']
+      @edit_api_limit.caption = setting['APILimit']     if setting['APILimit']
+      @tab_checkBox.check false                         if setting['TSV'] == false
       @author_edit.caption = setting['Author']          if setting['Author']
       @namecol_edit.caption = setting['SongnameColumn'] if setting['SongnameColumn']
       @hashcol_edit.caption = setting['SongIDcolumn']   if setting['SongIDcolumn']
-      @nameoutput_checkBox.check true                   if setting['SongnameOutput']
+      @nameoutput_checkBox.check false                  if setting['SongnameOutput'] == false
       self.x = setting['form_x']                        if setting['form_x']
       self.y = setting['form_y']                        if setting['form_y']
-    else
-      self.x = 330
-      self.y = 70
-      @tab_checkBox.check true
-      @nameoutput_checkBox.check true
+      if setting['sort_recent']
+        @radio_recent.check true
+        @radio_top.check false
+      end
     end
     self.caption += "  Ver #{SOFT_VER}"
     @prgrssBar1.setRange(0,100)
     @tz_static.caption = Time.now.zone
   end
   
-  #ScoreSaber Userデータ取得ボタン
-  def get_button_clicked
+  #ScoreSaber Userデータ プレイリスト変換ボタン
+  def get_button2_clicked
     #ボタンをDisableにする
     @get_button.style     = 1476395008
+    @get_button2.style    = 1476395008
     @imgopen_button.style = 1476395008
     @create_button.style  = 1476395008
     @playlist_tsv_button.style = 1476395008
     @prgrssBar1.position = 0
     refresh
     userid = @userid_edit.text.strip
-    limit_page = @limit_edit.text.strip.to_i
-    limit_page = false unless @limit_checkBox.checked? && limit_page > 0
-    player,rankchart,ss_data = user_score_get(userid,limit_page,@prgrssBar1,@read_static)
+    limit = @limit_edit.text.strip.to_i
+    limit = false unless @limit_checkBox.checked? && limit > 0
+    api_limit = @edit_api_limit.text.to_i
+    api_limit = 100 if api_limit <= 0
+    player,profile_picture,ss_data = user_score_get(userid,api_limit,limit,@radio_recent.checked?,@prgrssBar1,@read_static)
+    if ss_data
+      @prgrssBar1.position = 100
+      select_ext = '*.bplist'
+      file_type = ["BeatSaber PlayList(*.bplist)","*.bplist"]
+      if @radio_recent.checked?
+        sort = 'recent'
+      else
+        sort = 'top'
+      end
+      title = "#{player} #{sort} #{ss_data.size}"
+      playlist_data = playlist_convert2(title,profile_picture,ss_data)
+      if playlist_data
+        fn = SWin::CommonDialog::saveFilename(self,[file_type,["all file (*.*)","*.*"]],0x1004,'PLAYLIST SAVE FILE',select_ext,0,sjiscv(player + "_#{sort}").gsub(/[\\\/:\*\?\"<>|]/,'_'))
+        return unless fn
+        if File.exist?(fn)
+          return unless messageBox("Do you want to overwrite?","Overwrite confirmation",0x0004) == 6
+        end
+        File.open(fn,'w') do |file|
+          JSON.pretty_generate(playlist_data).each do |line|
+            file.puts line
+          end
+          puts "#{fn}:PlayList SaveOK!"
+        end
+      end
+    else
+      #取得データが異常時は警告表示
+      messageBox("ScoreSaber READ ERROR","ERROR",0)
+      @prgrssBar1.position = 0
+    end
+    #ボタンをEnableにする
+    @prgrssBar1.position = 0
+    @get_button.style     = 1342177280
+    @get_button2.style    = 1342177280
+    @imgopen_button.style = 1342177280
+    @create_button.style  = 1342177280
+    @playlist_tsv_button.style = 1342177280
+    @read_static.caption  = ""
+    refresh
+  end
+  
+  #ScoreSaber Userデータ TSV変換
+  def get_button_clicked
+    #ボタンをDisableにする
+    @get_button.style     = 1476395008
+    @get_button2.style    = 1476395008
+    @imgopen_button.style = 1476395008
+    @create_button.style  = 1476395008
+    @playlist_tsv_button.style = 1476395008
+    @prgrssBar1.position = 0
+    refresh
+    userid = @userid_edit.text.strip
+    limit = @limit_edit.text.strip.to_i
+    limit = false unless @limit_checkBox.checked? && limit > 0
+    api_limit = @edit_api_limit.text.to_i
+    api_limit = 100 if api_limit <= 0
+    player,profile_picture,ss_data = user_score_get(userid,api_limit,limit,@radio_recent.checked?,@prgrssBar1,@read_static)
     if @tab_checkBox.checked?
       separator = "\t"
       select_ext = '*.tsv'
@@ -236,7 +361,7 @@ module Frm_main_form
     end
     if ss_data
       @prgrssBar1.position = 100
-      fn = SWin::CommonDialog::saveFilename(self,[file_type,["all file (*.*)","*.*"]],0x1004,'USER SCORE LIST SAVE FILE',select_ext,0,"#{player}_scorelist")
+      fn = SWin::CommonDialog::saveFilename(self,[file_type,["all file (*.*)","*.*"]],0x1004,'USER SCORE LIST SAVE FILE',select_ext,0,"#{sjiscv(player)}_scorelist")
     else
       #取得データが異常時は警告表示
       messageBox("ScoreSaber READ ERROR","ERROR",0)
@@ -251,7 +376,7 @@ module Frm_main_form
       end
       if save_ok
         CSV.open(fn,'w',separator) do |record|
-          record << ['uid','id','rank','time','songBottom_time','pp','ppw','score','mapper','song_name']
+          record << ['uid','id','rank','time','pp','ppw','base score','modified score','max score','mapper','name','song_sub','difficulty']
           ss_data.each do |row|
             record << row
           end
@@ -261,6 +386,7 @@ module Frm_main_form
     #ボタンをEnableにする
     @prgrssBar1.position = 0
     @get_button.style     = 1342177280
+    @get_button2.style    = 1342177280
     @imgopen_button.style = 1342177280
     @create_button.style  = 1342177280
     @playlist_tsv_button.style = 1342177280
@@ -431,10 +557,11 @@ module Frm_main_form
     setting = {}
     setting['UserID']         = @userid_edit.text.strip
     if @limit_checkBox.checked?
-      setting['LimitPage']      = @limit_edit.text.strip
+      setting['Limit']      = @limit_edit.text.strip
     else
-      setting['LimitPage']      = false
+      setting['Limit']      = false
     end
+    setting['APILimit']       = @edit_api_limit.text.strip
     setting['TSV']            = @tab_checkBox.checked?
     setting['Author']         = @author_edit.text.strip
     setting['SongnameColumn'] = @namecol_edit.text.strip
@@ -442,6 +569,7 @@ module Frm_main_form
     setting['SongnameOutput'] = @nameoutput_checkBox.checked?
     setting['form_x']         = self.x
     setting['form_y']         = self.y
+    setting['sort_recent']    = @radio_recent.checked?
     File.open(SETTING_FILE,'w') do |file|
       JSON.pretty_generate(setting).each do |line|
         file.puts line
@@ -452,7 +580,8 @@ module Frm_main_form
   def default_menu_clicked
     @tab_checkBox.check true
     @nameoutput_checkBox.check true
-    @limit_checkBox.check false
+    @limit_checkBox.check true
+    @limit_edit.caption = 100
     @userid_edit.caption = ''
     @author_edit.caption = ''
     @namecol_edit.caption = 'name'
@@ -465,5 +594,45 @@ module Frm_main_form
 
 end
 
-#GUIメッセージループ
-VRLocalScreen.start Frm_main_form
+if ARGV == []
+  #GUIメッセージループ
+  VRLocalScreen.start Frm_main_form
+else
+  userid        = ARGV[0] #省略不可
+  playlist_file = ARGV[1] #省略不可
+  sort          = ARGV[2] #省略時 top
+  limit         = ARGV[3] #省略時 100
+  api_limit     = ARGV[4] #省略時 100
+  title         = ARGV[5] #省略時 プレイヤー ソート 取得数
+  if userid && playlist_file
+    limit = 100 unless limit
+    limit = limit.to_i
+    api_limit = 100 unless api_limit
+    api_limit = api_limit.to_i
+    api_limit = 100 if api_limit <= 0
+    sort_recent = false
+    sort_recent = true if sort =~ /recent/
+    if sort_recent
+      sort = "recent"
+    else
+      sort = "top"
+    end
+    player,profile_picture,ss_data = user_score_get(userid,api_limit,limit,sort_recent)
+    if ss_data
+      title = "#{player} #{sort} #{ss_data.size}" unless title
+      playlist_data = playlist_convert2(title,profile_picture,ss_data)
+      if playlist_data
+        File.open(playlist_file,'w') do |file|
+          JSON.pretty_generate(playlist_data).each do |line|
+            file.puts line
+          end
+          puts "#{playlist_file}:PlayList SaveOK!"
+        end
+      end
+    else
+      puts "User score data get error."
+    end
+  else
+    puts "userid and playlist file name are missing."
+  end
+end
